+++ src/System.Threading.AccessControl/ref/System.Threading.AccessControl.cs
@@ -19,6 +19,38 @@ namespace System.Security.AccessControl
         public EventWaitHandleAuditRule(System.Security.Principal.IdentityReference identity, System.Security.AccessControl.EventWaitHandleRights eventRights, System.Security.AccessControl.AuditFlags flags) : base(default(System.Security.Principal.IdentityReference), default(int), default(bool), default(System.Security.AccessControl.InheritanceFlags), default(System.Security.AccessControl.PropagationFlags), default(System.Security.AccessControl.AuditFlags)) { }
         public System.Security.AccessControl.EventWaitHandleRights EventWaitHandleRights { get { throw null; } }
     }
+    [System.FlagsAttribute]
+    public enum EventWaitHandleRights
+    {
+        ChangePermissions = 262144,
+        Delete = 65536,
+        FullControl = 2031619,
+        Modify = 2,
+        ReadPermissions = 131072,
+        Synchronize = 1048576,
+        TakeOwnership = 524288,
+    }
+
+    public sealed partial class EventWaitHandleSecurity : System.Security.AccessControl.NativeObjectSecurity
+    {
+        public EventWaitHandleSecurity() : base(default(bool), default(System.Security.AccessControl.ResourceType)) { }
+        public override System.Type AccessRightType { get { throw null; } }
+        public override System.Type AccessRuleType { get { throw null; } }
+        public override System.Type AuditRuleType { get { throw null; } }
+        public override System.Security.AccessControl.AccessRule AccessRuleFactory(System.Security.Principal.IdentityReference identityReference, int accessMask, bool isInherited, System.Security.AccessControl.InheritanceFlags inheritanceFlags, System.Security.AccessControl.PropagationFlags propagationFlags, System.Security.AccessControl.AccessControlType type) { throw null; }
+        public void AddAccessRule(System.Security.AccessControl.EventWaitHandleAccessRule rule) { }
+        public void AddAuditRule(System.Security.AccessControl.EventWaitHandleAuditRule rule) { }
+        public override System.Security.AccessControl.AuditRule AuditRuleFactory(System.Security.Principal.IdentityReference identityReference, int accessMask, bool isInherited, System.Security.AccessControl.InheritanceFlags inheritanceFlags, System.Security.AccessControl.PropagationFlags propagationFlags, System.Security.AccessControl.AuditFlags flags) { throw null; }
+        public bool RemoveAccessRule(System.Security.AccessControl.EventWaitHandleAccessRule rule) { throw null; }
+        public void RemoveAccessRuleAll(System.Security.AccessControl.EventWaitHandleAccessRule rule) { }
+        public void RemoveAccessRuleSpecific(System.Security.AccessControl.EventWaitHandleAccessRule rule) { }
+        public bool RemoveAuditRule(System.Security.AccessControl.EventWaitHandleAuditRule rule) { throw null; }
+        public void RemoveAuditRuleAll(System.Security.AccessControl.EventWaitHandleAuditRule rule) { }
+        public void RemoveAuditRuleSpecific(System.Security.AccessControl.EventWaitHandleAuditRule rule) { }
+        public void ResetAccessRule(System.Security.AccessControl.EventWaitHandleAccessRule rule) { }
+        public void SetAccessRule(System.Security.AccessControl.EventWaitHandleAccessRule rule) { }
+        public void SetAuditRule(System.Security.AccessControl.EventWaitHandleAuditRule rule) { }
+    }
     public sealed partial class MutexAccessRule : System.Security.AccessControl.AccessRule
     {
         public MutexAccessRule(System.Security.Principal.IdentityReference identity, System.Security.AccessControl.MutexRights eventRights, System.Security.AccessControl.AccessControlType type) : base(default(System.Security.Principal.IdentityReference), default(int), default(bool), default(System.Security.AccessControl.InheritanceFlags), default(System.Security.AccessControl.PropagationFlags), default(System.Security.AccessControl.AccessControlType)) { }
