+++ src/System.Threading.AccessControl/src/Configurations.props
@@ -1,14 +1,13 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<!-- Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information. -->
+<Project DefaultTargets="Build">
   <PropertyGroup>
     <PackageConfigurations>
-      net461-Windows_NT;
       netstandard-Windows_NT;
       netstandard;
     </PackageConfigurations>
     <BuildConfigurations>
       $(PackageConfigurations)
-      netfx-Windows_NT;
     </BuildConfigurations>
   </PropertyGroup>
 </Project>
