+++ src/System.Threading.AccessControl/tests/System.Threading.AccessControl.Tests.csproj
@@ -1,3 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE file in the project root for more information. -->
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <ProjectGuid>{458E445C-DF3C-4E4D-8E1D-F2FAC365BB40}</ProjectGuid>
